split sale_various into 2 parts
